organize types: CarrierType
