+++ types/Carrier/CarrierType/index.d.ts
@@ -1,0 +1,3 @@
+export * from './CarrierType';
+export * from './CarrierTypeFields';
+export * from './CarrierTypeCredentials';
