+++ types/Shipment/Shipment.d.ts
@@ -6,7 +6,8 @@ import { IFee } from '../Fee';
 import { IInsurance } from '../Insurance';
 import { IParcel } from '../Parcel';
 import { IScanForm } from '../ScanForm';
-import { Carrier, ITracker } from '../Tracker';
+import { Carrier } from '../Carrier';
+import { ITracker } from '../Tracker';
 import { IForm } from './Form';
 import { IMessage } from './Message';
 import { IOptions, LabelFormat } from './Options';
